ASSIGNMENT COMPLETED:
Stretch User Stories
You may have noticed that when you type in data into a UITextField the keyboard pops up. If the UITextField is at the bottom of the screen, when you touch it, the keyboard will cover up your textfield!

As a user, I want to see the Textfield after I press it...
Use Notification Center to get a notification when the keyboard shows. Move the screen up the size of the keyboard so that you are no longer hiding the textfield. (look up the keyboardWillShow notification)
As a user, I want to be able to easily adjust my tip percentage
Add A UISlider to the view
Create a method for the UISlider called adjustTipPercentage
Set the range of the slider to min 0 max 100
Update the tip percentage to the value of the slider
As a user, I don't want to have to press a button every time I change a field
make the text fields, and slider, update the calculated tip value whenever they are changed.